Code modified to introduce automation for MuonPeakPosition file generation, based on the information from the Database.
Database is also modified to contain 2 more flag [singlePointCalib & muonTrackFormation]
to check if the singlePointCalibration file is generated or not, and to check
if the muonTrack files are generated or not.

Code inside (test/toSchedule) directory is used to generated Muon Peak positins file.

Next steop would be to automate the MuonTrack Generation file.